tests/requirements-test.txt
# Testing dependencies for Omnishop

# Core testing framework
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0
pytest-timeout==2.2.0

# Test utilities
faker==20.1.0
factory-boy==3.3.0
freezegun==1.4.0

# HTTP testing
# Note: httpx version must be compatible with both google-genai (>=0.28.1) and pytest-httpx
# Using httpx 0.28.1 which is compatible with google-genai
# pytest-httpx tests use mocks so version incompatibility won't affect test execution
httpx>=0.28.1
pytest-httpx>=0.27.0

# Database testing
pytest-postgresql==5.0.0
sqlalchemy-utils==0.41.1

# Code quality
pytest-flake8==1.1.1
pytest-mypy==0.10.3
pytest-black==0.3.12

# Coverage reporting
coverage[toml]==7.3.2

# Mocking
responses==0.24.1
